backend/requirements.txt
# =============================================================================
# AI Video Editor SaaS - Backend Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- Database ---
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# --- Redis ---
redis==5.2.1

# --- Task Queue ---
celery[redis]==5.4.0

# --- Authentication & Security ---
argon2-cffi==23.1.0
python-multipart==0.0.19

# --- Validation & Config ---
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# --- Rate Limiting ---
slowapi==0.1.9

# --- Storage (S3) ---
boto3==1.35.92

# --- HTTP ---
httpx==0.28.1

# --- Utilities ---
python-dotenv==1.0.1

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
